wasm allocation test
Use `go run server.go` to run a webserver on localhost:8083 to test the
wasm build after building it.

It sets two headers which are required to allow the shared memory
feature for wasm files, without shared memory I wasn't able to get
allocations to work, maybe it is only supposed to work if shared-memory
is enabled.

The reason why shared-memory is disabled in browsers is because it
allows to create timing/side-channel attacks, that is why it needs to be
allowed for specific wasm files via cors settings.